% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ada_boost_svm_rbf.R
\name{ada_boost_svm_rbf}
\alias{ada_boost_svm_rbf}
\title{AdaBoost SVM with Radial Basis Function Kernel}
\usage{
ada_boost_svm_rbf(
  mode = "unknown",
  engine = "ebmc",
  num_learners = NULL,
  imb_ratio = NULL
)
}
\arguments{
\item{mode}{classification}

\item{engine}{ebmc's adam2 which uses e1701's svm}

\item{num_learners}{how many weak learners should be ensembled via boosting}

\item{imb_ratio}{major-minor class imbalance ratio}
}
\description{
AdaBoost SVM with Radial Basis Function Kernel
}
