% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apd_svm_novel_detection-fit.R
\name{apd_svm_novel_detection}
\alias{apd_svm_novel_detection}
\alias{apd_svm_novel_detection.default}
\alias{apd_svm_novel_detection.data.frame}
\alias{apd_svm_novel_detection.matrix}
\alias{apd_svm_novel_detection.formula}
\alias{apd_svm_novel_detection.recipe}
\title{Fit an SVM novelty detection model to estimate an applicability domain.}
\usage{
apd_svm_novel_detection(x, ...)

\method{apd_svm_novel_detection}{default}(x, ...)

\method{apd_svm_novel_detection}{data.frame}(x, ...)

\method{apd_svm_novel_detection}{matrix}(x, ...)

\method{apd_svm_novel_detection}{formula}(formula, data, ...)

\method{apd_svm_novel_detection}{recipe}(x, data, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}}

\item{...}{Options to pass to \code{\link[kernlab:ksvm]{kernlab::ksvm()}}. Options should
not include \code{data}.}

\item{formula}{A formula specifying the predictor terms on the right-hand
side. No outcome should be specified.}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing the predictors.
}}
}
\value{
A \code{apd_svm_novel_detection} object.
}
\description{
\code{apd_svm_novel_detection()} fits an 'one-svc' novelty detection model.
}
