% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcca_correlate.R
\name{autoplot.kcor_df}
\alias{autoplot.kcor_df}
\title{visualize component x-y feature space}
\usage{
\method{autoplot}{kcor_df}(x)
}
\arguments{
\item{x}{a kcor_df object}
}
\description{
visualize component x-y feature space
}
