% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{bessel_sigma}
\alias{bessel_sigma}
\alias{bessel_order}
\alias{laplace_sigma}
\title{Maize parameters}
\usage{
bessel_sigma(range = c(-10, 0), trans = scales::transform_log10())

bessel_order(range = c(1L, 3L), trans = NULL)

laplace_sigma(range = c(-10, 0), trans = scales::transform_log10())
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
Maize parameters
}
