% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_estimate_svm.R
\name{cal_estimate_svm}
\alias{cal_estimate_svm}
\alias{cal_estimate_svm.data.frame}
\alias{cal_estimate_svm.tune_results}
\alias{cal_estimate_svm.grouped_df}
\title{Uses a support vector regression model to calibrate numeric predictions}
\usage{
cal_estimate_svm(
  .data,
  truth = NULL,
  estimate = dplyr::matches("^.pred$"),
  smooth = TRUE,
  parameters = NULL,
  ...,
  .by = NULL
)

\method{cal_estimate_svm}{data.frame}(
  .data,
  truth = NULL,
  estimate = dplyr::matches("^.pred$"),
  smooth = TRUE,
  parameters = NULL,
  ...,
  .by = NULL
)

\method{cal_estimate_svm}{tune_results}(
  .data,
  truth = NULL,
  estimate = dplyr::matches("^.pred$"),
  smooth = TRUE,
  parameters = NULL,
  ...
)

\method{cal_estimate_svm}{grouped_df}(
  .data,
  truth = NULL,
  estimate = NULL,
  smooth = TRUE,
  parameters = NULL,
  ...
)
}
\arguments{
\item{.data}{An ungrouped  \code{data.frame} object, or \code{tune_results} object,
that contains a prediction column.}

\item{truth}{The column identifier for the observed outcome data (that is
numeric). This should be an unquoted column name.}

\item{estimate}{Column identifier for the predicted values}

\item{smooth}{Applies to the svm models. It switches between a polydot\code{TRUE},
and vanilladot \code{FALSE}.}

\item{parameters}{(Optional)  An optional tibble of tuning parameter values
that can be used to filter the predicted values before processing. Applies
only to \code{tune_results} objects.}

\item{...}{Additional arguments passed to the models or routines used to
calculate the new predictions.}
}
\description{
Uses a support vector regression model to calibrate numeric predictions
}
\details{
This function uses existing modeling functions from other packages to create
the calibration:
\itemize{
\item \code{\link[kernlab:ksvm]{kernlab::ksvm()}} with a "vanilladot" is used when \code{smooth} is set to \code{FALSE}
\item \code{\link[kernlab:ksvm]{kernlab::ksvm()}} with a "polydot" is used when \code{smooth} is set to \code{TRUE}
}

These methods estimate the relationship in the unmodified predicted values
and then remove that trend when \code{\link[=cal_apply]{cal_apply()}} is invoked.
}
\seealso{
\url{https://www.tidymodels.org/learn/models/calibration/},
}
