% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corn_data.R
\docType{data}
\name{corn_data}
\alias{corn_data}
\title{Synthetic Corn Dataset for Corny Example}
\source{
claude-3-5-sonnet-20240620
}
\value{
\item{corn_data}{a tibble}
}
\description{
Synthetic Corn Dataset for Corny Example
}
\details{
Asked Claude Sonnet for a corn data given the README story problem
}
\examples{
data(corn_data)
str(corn_data)
}
\keyword{datasets}
