% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_conformal_quantile_svm.R
\name{int_conformal_quantile_svm}
\alias{int_conformal_quantile_svm}
\alias{int_conformal_quantile_svm.workflow}
\title{Prediction intervals via conformal inference and QRSVM}
\usage{
int_conformal_quantile_svm(object, ...)

\method{int_conformal_quantile_svm}{workflow}(object, train_data, cal_data, level = 0.95, ...)
}
\arguments{
\item{object}{A fitted \code{\link[workflows:workflow]{workflows::workflow()}} object.}

\item{...}{Options to pass to \code{\link[qrsvm:qrsvm]{qrsvm::qrsvm()}}}

\item{train_data, cal_data}{Data frames with the \emph{predictor and outcome data}.
\code{train_data} should be the same data used to produce \code{object} and \code{cal_data} is
used to produce predictions (and residuals). If the workflow used a recipe,
these should be the data that were inputs to the recipe (and not the product
of a recipe).}

\item{level}{The confidence level for the intervals.}
}
\value{
An object of class \code{"int_conformal_quantile"} containing the
information to create intervals (which includes \code{object}).
The \code{predict()} method is used to produce the intervals.
}
\description{
To compute quantiles, this function uses Quantile
SVM instead of probably's \code{"int_conformal_quantile"} QRF or classic quantile regression.
}
\details{
based on the initial probably implementation with slight modification.

origin: https://github.com/tidymodels/probably/blob/HEAD/R/conformal_infer_quantile.R

for more information, visit:
https://probably.tidymodels.org
}
