% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcca_correlate.R
\name{kcca_correlate}
\alias{kcca_correlate}
\title{Kernel Canonical Correlation Analysis}
\usage{
kcca_correlate(x, y = NULL, kernel = "rbfdot", gamma = 0.1, num_comp = 10, ...)
}
\arguments{
\item{x}{variable or dataframe}

\item{y}{variable or dataframe}

\item{kernel}{a kernel to use}

\item{gamma}{regularization parameter}

\item{num_comp}{number of components}

\item{...}{pass through args for kcca function}
}
\value{
A kernel canonical correlation analysis data frame \code{kcor_df}
}
\description{
Computes the canonical correlation analysis in feature space.
Kernel Canonical Correlation Analysis (KCCA) is a non-linear extension of CCA.
Given two random variables (or datasets), KCCA aims at extracting the information which is shared by the two random variables (or datasets).
more information found at \code{\link[kernlab:kcca]{kernlab::kcca()}}
}
