% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kqr_laplace.R
\name{kqr_laplace}
\alias{kqr_laplace}
\title{Laplacian Kernel Quantile Regression}
\usage{
kqr_laplace(
  mode = "unknown",
  engine = "kernlab",
  cost = NULL,
  tau = NULL,
  laplace_sigma = NULL
)
}
\arguments{
\item{mode}{regression or classification}

\item{engine}{kernlab kqr}

\item{cost}{A positive number for the cost of predicting a sample within
or on the wrong side of the margin}

\item{tau}{a quantile for the loss function}

\item{laplace_sigma}{sigma parameter for laplacian}
}
\description{
laplacian kernel
}
