% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lssvm_laplace.R
\name{lssvm_laplace}
\alias{lssvm_laplace}
\title{Laplacian Least Squares Support Vector Machine}
\usage{
lssvm_laplace(mode = "unknown", engine = "kernlab", laplace_sigma = NULL)
}
\arguments{
\item{mode}{regression or classification}

\item{engine}{kernlab lssvm}

\item{laplace_sigma}{sigma parameter for laplacian}
}
\description{
laplacian kernel
}
