% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maize-package.R
\docType{package}
\name{maize-package}
\alias{maize}
\alias{maize-package}
\title{maize: Specialty Kernels for SVMs}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Bindings for svm kernels via kernlab for use with the 'parsnip' package. Specifically related to specialty kernels for support vector machines not available in parsnip. package includes interface for various kernlab kernels and custom kernels too.
}
\author{
\strong{Maintainer}: Frankie T. Hull \email{frankiethull@gmail.com}

Other contributors:
\itemize{
  \item Max Kuhn \email{max@posit.co} [contributor]
}

}
\keyword{internal}
