% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misvm_orova_rbf.R
\name{misvm_orova_rbf}
\alias{misvm_orova_rbf}
\title{Multiple Instance SVM Ordinal Outcome Data One-Vs-All with Radial Basis Function Kernel}
\usage{
misvm_orova_rbf(
  mode = "unknown",
  engine = "ebmc",
  cost = NULL,
  rbf_sigma = NULL
)
}
\arguments{
\item{mode}{classification}

\item{engine}{\code{mildsvm::misvm_orova()} which uses e1701's svm}

\item{cost}{A positive number for the cost of predicting a sample within
or on the wrong side of the margin}

\item{rbf_sigma}{A positive number for radial basis function.}
}
\description{
Multiple Instance SVM Ordinal Outcome Data One-Vs-All with Radial Basis Function Kernel
}
