% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rus_boost_svm_rbf.R
\name{rus_boost_svm_rbf}
\alias{rus_boost_svm_rbf}
\title{RUSBoost SVM with Radial Basis Function Kernel}
\usage{
rus_boost_svm_rbf(
  mode = "unknown",
  engine = "ebmc",
  num_learners = NULL,
  imb_ratio = NULL
)
}
\arguments{
\item{mode}{classification}

\item{engine}{ebmc's rus which uses e1701's svm}

\item{num_learners}{how many weak learners should be ensembled via boosting}

\item{imb_ratio}{major-minor class imbalance ratio}
}
\description{
RUSBoost SVM with Radial Basis Function Kernel
}
