% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apd_svm_novel_detection-score.R
\name{score.apd_svm_novel_detection}
\alias{score.apd_svm_novel_detection}
\title{Predict from a \code{apd_svm_novel_detection}}
\usage{
\method{score}{apd_svm_novel_detection}(object, new_data, type = "numeric", ...)
}
\arguments{
\item{object}{A \code{apd_svm_novel_detection} object.}

\item{new_data}{A data frame or matrix of new samples.}

\item{type}{A single character. The type of predictions to generate.
Valid options are:
\itemize{
\item \code{"numeric"} for numeric predictions.
}}

\item{...}{Not used, but required for extensibility.}
}
\value{
A tibble of predictions. The number of rows in the tibble is guaranteed
to be the same as the number of rows in \code{new_data}. The \code{score} column is the
raw prediction from \code{\link[kernlab:predict]{kernlab::predict()}} while \code{score_pctl}
compares this value to the reference distribution of the score created by
predicting the training set. A value of \emph{X} means that \emph{X} percent of the
training data have scores less than the predicted value.
}
\description{
Predict from a \code{apd_svm_novel_detection}
}
\details{
About the score
}
\seealso{
\code{\link[=apd_svm_novel_detection]{apd_svm_novel_detection()}}
}
