% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_kfm_nystrom.R
\name{step_kfm_nystrom}
\alias{step_kfm_nystrom}
\title{Nystrom kernel feature map approximation (RBF)}
\usage{
step_kfm_nystrom(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  res = NULL,
  columns = NULL,
  sigma = 0.2,
  m = NULL,
  r = NULL,
  sampling = "random",
  prefix = "kFM",
  keep_original_cols = FALSE,
  skip = FALSE,
  id = rand_id("kfm_nystrom")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose variables
for this step. See \code{\link[=selections]{selections()}} for more details.}

\item{role}{For model terms created by this step, what analysis role should
they be assigned? By default, the new columns created by this step from
the original variables will be used as \emph{predictors} in a model.}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{res}{An \code{\link[mildsvm:kfm_nystrom]{mildsvm::kfm_nystrom()}} object is stored
here once this preprocessing step has be trained by
\code{\link[=prep]{prep()}}.}

\item{columns}{A character string of the selected variable names. This field
is a placeholder and will be populated once \code{\link[=prep]{prep()}} is used.}

\item{sigma}{A numeric value for the nystrom function parameter.}

\item{m}{The number rows from df to sample in fitting. defaults to nrow of data}

\item{r}{The rank of matrix approximation to use. Must be less than or equal to m, the default.}

\item{prefix}{A character string for the prefix of the resulting new
variables. See notes below.}

\item{keep_original_cols}{A logical to keep the original variables in the
output. Defaults to \code{FALSE}.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[=bake]{bake()}}? While all operations are baked
when \code{\link[=prep]{prep()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}
}
\description{
\code{step_kfm_nystrom()} creates a \emph{specification} of a recipe step that will
convert numeric data into a feature appoximation. nystrom approximates the 'radial'
kernel approximation.
}
\seealso{
Other multivariate transformation steps: 
\code{\link{step_kcca_laplace}()},
\code{\link{step_kfa_laplace}()},
\code{\link{step_kha_laplace}()},
\code{\link{step_kha_tanh}()},
\code{\link{step_kpca_laplace}()},
\code{\link{step_kpca_tanh}()}
}
\concept{multivariate transformation steps}
