% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_kpca_laplace.R
\name{step_kpca_laplace}
\alias{step_kpca_laplace}
\title{Laplacian function kernel PCA signal extraction}
\usage{
step_kpca_laplace(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  num_comp = 5,
  res = NULL,
  columns = NULL,
  sigma = 0.2,
  prefix = "kPC",
  keep_original_cols = FALSE,
  skip = FALSE,
  id = rand_id("kpca_laplace")
)
}
\arguments{
\item{res}{An S4 \code{\link[kernlab:kpca]{kernlab::kpca()}} object is stored
here once this preprocessing step has be trained by
\code{\link[=prep]{prep()}}.}

\item{sigma}{A numeric value for the laplace function parameter.}
}
\description{
\code{step_kpca_laplace()} creates a \emph{specification} of a recipe step that will
convert numeric data into one or more principal components using a laplace
kernel basis expansion.
}
\seealso{
Other multivariate transformation steps: 
\code{\link{step_kha_laplace}()}
}
\concept{multivariate transformation steps}
