% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_anova_rbf.R
\name{svm_anova_rbf}
\alias{svm_anova_rbf}
\title{ANOVA RBF Support Vector Machine}
\usage{
svm_anova_rbf(
  mode = "unknown",
  engine = "kernlab",
  cost = NULL,
  anova_rbf_sigma = NULL,
  degree = NULL,
  margin = NULL
)
}
\arguments{
\item{mode}{regression or classification}

\item{engine}{kernlab ksvm}

\item{cost}{A positive number for the cost of predicting a sample within
or on the wrong side of the margin}

\item{anova_rbf_sigma}{sigma parameter for anova rbf}

\item{degree}{degree parameter for anova rbf}

\item{margin}{A positive number for the epsilon in the SVM insensitive
loss function (regression only)}
}
\description{
anova rbf for support vector machines
}
