% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_bessel.R
\name{svm_bessel}
\alias{svm_bessel}
\title{Bessel Support Vector Machine}
\usage{
svm_bessel(
  mode = "unknown",
  engine = "kernlab",
  cost = NULL,
  bessel_sigma = NULL,
  degree = NULL,
  order = NULL,
  margin = NULL
)
}
\arguments{
\item{mode}{regression or classification}

\item{engine}{kernlab ksvm}

\item{cost}{A positive number for the cost of predicting a sample within
or on the wrong side of the margin}

\item{bessel_sigma}{sigma parameter for bessel}

\item{degree}{degree parameter for bessel}

\item{order}{order parameter for bessel}

\item{margin}{A positive number for the epsilon in the SVM insensitive
loss function (regression only)}
}
\description{
bessel kernel for support vector machines
}
