% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_cossim.R
\name{svm_cossim}
\alias{svm_cossim}
\title{Cosine Similarity Support Vector Machine}
\usage{
svm_cossim(mode = "unknown", engine = "kernlab", cost = NULL, margin = NULL)
}
\arguments{
\item{mode}{regression or classification}

\item{engine}{kernlab ksvm}

\item{cost}{A positive number for the cost of predicting a sample within
or on the wrong side of the margin}

\item{margin}{A positive number for the epsilon in the SVM insensitive
loss function (regression only)}
}
\description{
cossim kernel for support vector machines
}
