% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_laplace.R
\name{svm_laplace}
\alias{svm_laplace}
\title{Laplacian Support Vector Machine}
\usage{
svm_laplace(
  mode = "unknown",
  engine = "kernlab",
  cost = NULL,
  margin = NULL,
  laplace_sigma = NULL
)
}
\arguments{
\item{mode}{regression or classification}

\item{engine}{kernlab ksvm}

\item{cost}{A positive number for the cost of predicting a sample within
or on the wrong side of the margin}

\item{margin}{A positive number for the epsilon in the SVM insensitive
loss function (regression only)}

\item{laplace_sigma}{sigma parameter for laplacian}
}
\description{
laplacian kernel for support vector machines
}
