% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_sorensen.R
\name{svm_sorensen}
\alias{svm_sorensen}
\title{Sorensen Support Vector Machine}
\usage{
svm_sorensen(mode = "unknown", engine = "kernlab", cost = NULL, margin = NULL)
}
\arguments{
\item{mode}{regression or classification}

\item{engine}{kernlab ksvm}

\item{cost}{A positive number for the cost of predicting a sample within
or on the wrong side of the margin}

\item{margin}{A positive number for the epsilon in the SVM insensitive
loss function (regression only)}
}
\description{
sorensen kernel for support vector machines which is used as a graph kernel for chemical informatics
}
