% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_string.R
\name{svm_string}
\alias{svm_string}
\title{String Support Vector Machine}
\usage{
svm_string(
  x,
  y,
  mode = "unknown",
  engine = "kernlab",
  cost = NULL,
  length = NULL,
  lambda = NULL,
  normalized = TRUE,
  margin = NULL
)
}
\arguments{
\item{mode}{regression or classification}

\item{engine}{kernlab ksvm}

\item{cost}{A positive number for the cost of predicting a sample within
or on the wrong side of the margin}

\item{length}{The length of the substrings considered}

\item{lambda}{The decay factor}

\item{normalized}{normalize string kernel values, (default = TRUE)}

\item{margin}{A positive number for the epsilon in the SVM insensitive
loss function (regression only)}
}
\description{
stringdot for support vector machines
}
