% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_tanh.R
\name{svm_tanh}
\alias{svm_tanh}
\title{Hyperbolic Tangent Support Vector Machine}
\usage{
svm_tanh(
  mode = "unknown",
  engine = "kernlab",
  cost = NULL,
  scale_factor = NULL,
  margin = NULL
)
}
\arguments{
\item{mode}{regression or classification}

\item{engine}{kernlab ksvm}

\item{cost}{A positive number for the cost of predicting a sample within
or on the wrong side of the margin}

\item{scale_factor}{scale parameter for tanh}

\item{margin}{A positive number for the epsilon in the SVM insensitive
loss function (regression only)}
}
\description{
tanh kernel for support vector machines
}
