% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_tanimoto.R
\name{svm_tanimoto}
\alias{svm_tanimoto}
\title{Tanimoto Support Vector Machine}
\usage{
svm_tanimoto(mode = "unknown", engine = "kernlab", cost = NULL, margin = NULL)
}
\arguments{
\item{mode}{regression or classification}

\item{engine}{kernlab ksvm}

\item{cost}{A positive number for the cost of predicting a sample within
or on the wrong side of the margin}

\item{margin}{A positive number for the epsilon in the SVM insensitive
loss function (regression only)}
}
\description{
tanimoto kernel for support vector machines which is used as a graph kernel for chemical informatics
}
