% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_tstudent.R
\name{svm_tstudent}
\alias{svm_tstudent}
\title{T-Student Support Vector Machine}
\usage{
svm_tstudent(
  mode = "unknown",
  engine = "kernlab",
  cost = NULL,
  margin = NULL,
  degree = NULL
)
}
\arguments{
\item{mode}{regression or classification}

\item{engine}{kernlab ksvm}

\item{cost}{A positive number for the cost of predicting a sample within
or on the wrong side of the margin}

\item{margin}{A positive number for the epsilon in the SVM insensitive
loss function (regression only)}

\item{degree}{a degree parameter for tstudent kernels}
}
\description{
t-student kernel for support vector machines
}
