% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maize-package.R
\docType{package}
\name{maize-package}
\alias{maize}
\alias{maize-package}
\title{maize: Specialty Kernels for SVMs}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Bindings for SVMs and special kernels via kernlab, e1701, ebmc, and mildsvm extending the 'parsnip' package. Specifically related to kernels for support vector machines not available in parsnip. Package also includes interfaces for novel SVM preprocessors (recipes) and postprocessors (probably).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://frankiethull.github.io/maize/}
}

}
\author{
\strong{Maintainer}: Frank Hull \email{frankiethull@gmail.com}

Other contributors:
\itemize{
  \item Max Kuhn \email{max@posit.co} [contributor]
}

}
\keyword{internal}
