% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_kha_laplace.R
\name{step_kha_laplace}
\alias{step_kha_laplace}
\title{Laplacian function kernel PCA signal extraction via Hebbian Algorithm}
\usage{
step_kha_laplace(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  num_comp = 5,
  res = NULL,
  columns = NULL,
  sigma = 0.2,
  learn_rate = 0.005,
  threshold = 1e-04,
  stop_iter = 100,
  prefix = "kha",
  keep_original_cols = FALSE,
  skip = FALSE,
  id = rand_id("kha_laplace")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose variables
for this step. See \code{\link[=selections]{selections()}} for more details.}

\item{role}{For model terms created by this step, what analysis role should
they be assigned? By default, the new columns created by this step from
the original variables will be used as \emph{predictors} in a model.}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{num_comp}{The number of components to retain as new predictors.
If \code{num_comp} is greater than the number of columns or the number of
possible components, a smaller value will be used. If \code{num_comp = 0}
is set then no transformation is done and selected variables will
stay unchanged, regardless of the value of \code{keep_original_cols}.}

\item{res}{An S4 \code{\link[kernlab:kha]{kernlab::kha()}} object is stored
here once this preprocessing step has be trained by
\code{\link[=prep]{prep()}}.}

\item{columns}{A character string of the selected variable names. This field
is a placeholder and will be populated once \code{\link[=prep]{prep()}} is used.}

\item{sigma}{A numeric value for the laplace function parameter.}

\item{learn_rate}{hebbian learning rate}

\item{threshold}{the smallest value of the convergence step}

\item{stop_iter}{maximum number of iterations}

\item{prefix}{A character string for the prefix of the resulting new
variables. See notes below.}

\item{keep_original_cols}{A logical to keep the original variables in the
output. Defaults to \code{FALSE}.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[=bake]{bake()}}? While all operations are baked
when \code{\link[=prep]{prep()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}
}
\description{
\code{step_kha_laplace()} creates a \emph{specification} of a recipe step that will
convert numeric data into one or more principal components using a laplace
kernel basis expansion.
}
\seealso{
Other multivariate transformation steps: 
\code{\link{step_kfa_laplace}()},
\code{\link{step_kfm_nystrom}()},
\code{\link{step_kha_tanh}()},
\code{\link{step_kpca_laplace}()},
\code{\link{step_kpca_tanh}()}
}
\concept{multivariate transformation steps}
